pyproject.toml
[tool.poetry]
name = "kek-cheburek"
version = "0.1.0"
description = ""
authors = ["Drvmnekta <nataliakamysheva@mail.ru>"]
readme = "README.md"
packages = [{include = "main.py"}]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^9.2.0"
pynput = "^1.7.6"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
